Secure profile and product admin views

+++ products/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, reverse, redirect, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductForm
 
@@ -60,10 +61,17 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     A view to add a product to the store
     """
+    # Check user has permission to access the function
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission \
+            to access this function.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -85,10 +93,17 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     A view to edit a product
     """
+    # Check user has permission to access the function
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission \
+            to access this function.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
 
     if request.method == 'POST':
@@ -113,10 +128,17 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product from the store
     """
+    # Check user has permission to access the function
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have permission \
+            to access this function.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product has been deleted.')
